Change directories of some packages
