+++ sources/mdtk/mdtk.go
@@ -7,8 +7,8 @@ import (
 	"mdtk/parse"
 	"mdtk/help"
 	"mdtk/path"
-	"mdtk/read"
-	"mdtk/grtask"
+	"mdtk/taskset/read"
+	"mdtk/taskset/grtask"
 	"mdtk/taskset"
 	"mdtk/args"
 	"mdtk/cache"
