+++ sources/mdtk/taskset/code/code_embed_config_once.go
@@ -3,7 +3,7 @@ package code
 import (
 	"strings"
 	"mdtk/lib"
-	"mdtk/grtask"
+	"mdtk/taskset/grtask"
 )
 
 type TaskStackData struct {
