+++ sources/mdtk/taskset/code/code_embed_desc.go
