+++ sources/mdtk/taskset/read/read_task.go
@@ -5,9 +5,9 @@ import (
 	"strings"
 	"regexp"
 	"mdtk/base"
-	"mdtk/group"
-	"mdtk/task"
-	"mdtk/code"
+	"mdtk/taskset/group"
+	"mdtk/taskset/task"
+	"mdtk/taskset/code"
 	"mdtk/taskset"
 	"mdtk/path"
 	"mdtk/config"
