+++ sources/mdtk/taskset/read/read_taskconfig_base.go
