+++ sources/mdtk/taskset/taskdata.go
@@ -2,9 +2,9 @@ package taskset
 
 import (
 	"strings"
-	"mdtk/group"
-	"mdtk/task"
-	"mdtk/code"
+	"mdtk/taskset/group"
+	"mdtk/taskset/task"
+	"mdtk/taskset/code"
 	"mdtk/path"
 )
